Separate out svc and irc bits (See description)
Broke things out into regular functions for easier re-use. This also
separates things that are specific to irc and things that are specific
to the service. Things that are specific to irc are prefixed with irc_
and things that are specific to the service are prefixed svc_. Some of
the svc_ functions expect a MessageService object, but it doesn't
matter if the object is backed with irc or some other backend.

~ B'ezrat Hashem ~